Move mapNotification to a util
